requirements.txt
# django>=3.0,<4.0
django>=4.2.16
djangorestframework
psycopg2
pytest
pytest-django
pytest-env
factory_boy
pytest-cov
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
# Django==5.1.3
django-ckeditor-5==0.2.15
django-ckeditor5==0.0.4
django-js-asset==2.2.0
django-shortuuidfield==0.1.3
fonttools==4.55.0
kiwisolver==1.4.7
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
# django-allauth
# requests

asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
# Django==5.1.3
django-allauth==65.3.0
# django-ckeditor-5==0.2.15
# django-shortuuidfield==0.1.3
idna==3.10
# pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
# shortuuid==1.0.13
# six==1.16.0
# sqlparse==0.5.2
# tzdata==2024.2
urllib3==2.2.3

dj-database-url
# python-dotenv==1.0.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==24.2
psycopg==3.2.3
psycopg-pool==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
python-dotenv==1.0.1
service-identity==24.2.0
setuptools==75.6.0
sqlparse==0.5.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
zope.interface==7.2

channels